+++ msi.gama.core/src/msi/gaml/types/GamaMatrixType.java
@@ -1,8 +1,7 @@
 /*********************************************************************************************
  *
- * 'GamaMatrixType.java, in plugin msi.gama.core, is part of the source code of the
- * GAMA modeling and simulation platform.
- * (c) 2007-2016 UMI 209 UMMISCO IRD/UPMC & Partners
+ * 'GamaMatrixType.java, in plugin msi.gama.core, is part of the source code of the GAMA modeling and simulation
+ * platform. (c) 2007-2016 UMI 209 UMMISCO IRD/UPMC & Partners
  *
  * Visit https://github.com/gama-platform/gama for license information and developers contact.
  * 
@@ -33,31 +32,27 @@ import msi.gaml.expressions.ListExpression;
 import msi.gaml.expressions.MapExpression;
 import msi.gaml.operators.Cast;
 
-@type(name = IKeyword.MATRIX, id = IType.MATRIX, wraps = { IMatrix.class, GamaIntMatrix.class, GamaFloatMatrix.class,
-		GamaObjectMatrix.class }, kind = ISymbolKind.Variable.CONTAINER, concept = { IConcept.TYPE, IConcept.CONTAINER,
-				IConcept.MATRIX })
-@SuppressWarnings({ "unchecked", "rawtypes" })
+@type (
+		name = IKeyword.MATRIX,
+		id = IType.MATRIX,
+		wraps = { IMatrix.class, GamaIntMatrix.class, GamaFloatMatrix.class, GamaObjectMatrix.class },
+		kind = ISymbolKind.Variable.CONTAINER,
+		concept = { IConcept.TYPE, IConcept.CONTAINER, IConcept.MATRIX })
+@SuppressWarnings ({ "unchecked", "rawtypes" })
 public class GamaMatrixType extends GamaContainerType<IMatrix> {
 
 	public static IMatrix staticCast(final IScope scope, final Object obj, final Object param, final IType contentType,
 			final boolean copy) {
-		if (obj == null && param == null) {
-			return null;
-		}
+		if (obj == null && param == null) { return null; }
 		final GamaPoint size = param instanceof GamaPoint ? (GamaPoint) param : null;
 
 		if (size == null) {
-			if (obj instanceof IContainer) {
-				return ((IContainer) obj).matrixValue(scope, contentType, copy);
-			}
+			if (obj instanceof IContainer) { return ((IContainer) obj).matrixValue(scope, contentType, copy); }
 			return with(scope, obj, new GamaPoint(1, 1), contentType);
-		} else if (size.x <= 0 || size.y < 0) {
-			throw GamaRuntimeException.error("Dimensions of a matrix should be positive.", scope);
-		}
+		} else if (size.x <= 0 || size.y < 0) { throw GamaRuntimeException
+				.error("Dimensions of a matrix should be positive.", scope); }
 
-		if (obj instanceof IContainer) {
-			return ((IContainer) obj).matrixValue(scope, contentType, size, copy);
-		}
+		if (obj instanceof IContainer) { return ((IContainer) obj).matrixValue(scope, contentType, size, copy); }
 		return with(scope, obj, size, contentType);
 
 	}
@@ -71,9 +66,7 @@ public class GamaMatrixType extends GamaContainerType<IMatrix> {
 
 	public static IMatrix from(final IScope scope, final IList list, final IType desiredType,
 			final ILocation preferredSize) {
-		if (list == null || list.isEmpty()) {
-			return new GamaObjectMatrix(0, 0, desiredType);
-		}
+		if (list == null || list.isEmpty()) { return new GamaObjectMatrix(0, 0, desiredType); }
 		if (desiredType.id() == IType.INT) {
 			return new GamaIntMatrix(scope, list, preferredSize);
 		} else if (desiredType.id() == IType.FLOAT) {
@@ -100,9 +93,7 @@ public class GamaMatrixType extends GamaContainerType<IMatrix> {
 	public static IMatrix from(final IScope scope, final IMatrix matrix, final IType desiredType,
 			final ILocation preferredSize, final boolean copy) {
 		final IType contentsType = matrix.getType().getContentType();
-		if (!GamaType.requiresCasting(desiredType, contentsType)) {
-			return matrix.copy(scope, preferredSize, copy);
-		}
+		if (!GamaType.requiresCasting(desiredType, contentsType)) { return matrix.copy(scope, preferredSize, copy); }
 		int cols, rows;
 		if (preferredSize == null) {
 			cols = matrix.getCols(scope);
@@ -112,17 +103,17 @@ public class GamaMatrixType extends GamaContainerType<IMatrix> {
 			rows = (int) preferredSize.getY();
 		}
 		switch (desiredType.id()) {
-		case IType.INT:
-			return GamaIntMatrix.from(scope, cols, rows, matrix);
-		case IType.FLOAT:
-			return GamaFloatMatrix.from(scope, cols, rows, matrix);
-		default:
-			final GamaObjectMatrix m = GamaObjectMatrix.from(cols, rows, matrix);
-			final Object[] array = m.getMatrix();
-			for (int i = 0; i < array.length; i++) {
-				array[i] = desiredType.cast(scope, array[i], null, false);
-			}
-			return m;
+			case IType.INT:
+				return GamaIntMatrix.from(scope, cols, rows, matrix);
+			case IType.FLOAT:
+				return GamaFloatMatrix.from(scope, cols, rows, matrix);
+			default:
+				final GamaObjectMatrix m = GamaObjectMatrix.from(cols, rows, matrix);
+				final Object[] array = m.getMatrix();
+				for (int i = 0; i < array.length; i++) {
+					array[i] = desiredType.cast(scope, array[i], null, false);
+				}
+				return m;
 		}
 
 	}
@@ -134,39 +125,38 @@ public class GamaMatrixType extends GamaContainerType<IMatrix> {
 
 	public static IMatrix with(final IScope scope, final IExpression fillExpr, final int cols, final int rows) {
 		IMatrix result;
-		if (fillExpr == null) {
-			return new GamaObjectMatrix(cols, rows, Types.NO_TYPE);
-		}
+		if (fillExpr == null) { return new GamaObjectMatrix(cols, rows, Types.NO_TYPE); }
 		switch (fillExpr.getType().id()) {
-		case IType.FLOAT:
-			result = new GamaFloatMatrix(cols, rows);
-			final double[] dd = ((GamaFloatMatrix) result).getMatrix();
-			if (fillExpr.isConst()) {
-				Arrays.fill(dd, Cast.asFloat(scope, fillExpr.value(scope)));
-			} else
-				GamaExecutorService.executeThreaded(() -> IntStream.range(0, dd.length).parallel().forEach(i -> {
-					dd[i] = Cast.asFloat(scope, fillExpr.value(scope));
-				}));
-			break;
-		case IType.INT:
-			result = new GamaIntMatrix(cols, rows);
-			final int[] ii = ((GamaIntMatrix) result).getMatrix();
-			if (fillExpr.isConst()) {
-				Arrays.fill(ii, Cast.asInt(scope, fillExpr.value(scope)));
-			} else
-				GamaExecutorService.executeThreaded(() -> IntStream.range(0, ii.length).parallel().forEach(i -> {
-					ii[i] = Cast.asInt(scope, fillExpr.value(scope));
-				}));
-			break;
-		default:
-			result = new GamaObjectMatrix(cols, rows, fillExpr.getType());
-			final Object[] contents = ((GamaObjectMatrix) result).getMatrix();
-			if (fillExpr.isConst()) {
-				Arrays.fill(contents, fillExpr.value(scope));
-			} else
-				GamaExecutorService.executeThreaded(() -> IntStream.range(0, contents.length).parallel().forEach(i -> {
-					contents[i] = fillExpr.value(scope);
-				}));
+			case IType.FLOAT:
+				result = new GamaFloatMatrix(cols, rows);
+				final double[] dd = ((GamaFloatMatrix) result).getMatrix();
+				if (fillExpr.isConst()) {
+					Arrays.fill(dd, Cast.asFloat(scope, fillExpr.value(scope)));
+				} else
+					GamaExecutorService.executeThreaded(() -> IntStream.range(0, dd.length).parallel().forEach(i -> {
+						dd[i] = Cast.asFloat(scope, fillExpr.value(scope));
+					}));
+				break;
+			case IType.INT:
+				result = new GamaIntMatrix(cols, rows);
+				final int[] ii = ((GamaIntMatrix) result).getMatrix();
+				if (fillExpr.isConst()) {
+					Arrays.fill(ii, Cast.asInt(scope, fillExpr.value(scope)));
+				} else
+					GamaExecutorService.executeThreaded(() -> IntStream.range(0, ii.length).parallel().forEach(i -> {
+						ii[i] = Cast.asInt(scope, fillExpr.value(scope));
+					}));
+				break;
+			default:
+				result = new GamaObjectMatrix(cols, rows, fillExpr.getType());
+				final Object[] contents = ((GamaObjectMatrix) result).getMatrix();
+				if (fillExpr.isConst()) {
+					Arrays.fill(contents, fillExpr.value(scope));
+				} else
+					GamaExecutorService
+							.executeThreaded(() -> IntStream.range(0, contents.length).parallel().forEach(i -> {
+								contents[i] = fillExpr.value(scope);
+							}));
 		}
 		return result;
 	}
@@ -180,11 +170,11 @@ public class GamaMatrixType extends GamaContainerType<IMatrix> {
 			final IType contentsType) throws GamaRuntimeException {
 		if (contentsType == Types.INT || val instanceof Integer) {
 			final GamaIntMatrix matrix = new GamaIntMatrix(cols, rows);
-			matrix.setAllValues(scope, (Integer) Types.INT.cast(scope, val, null, false));
+			matrix.setAllValues(scope, Types.INT.cast(scope, val, null, false));
 			return matrix;
 		} else if (contentsType == Types.FLOAT || val instanceof Double) {
 			final GamaFloatMatrix matrix = new GamaFloatMatrix(cols, rows);
-			matrix.setAllValues(scope, (Double) Types.FLOAT.cast(scope, val, null, false));
+			matrix.setAllValues(scope, Types.FLOAT.cast(scope, val, null, false));
 			return matrix;
 		}
 		final IMatrix matrix = new GamaObjectMatrix(cols, rows, contentsType);
@@ -209,23 +199,17 @@ public class GamaMatrixType extends GamaContainerType<IMatrix> {
 		if (itemType.id() == IType.LIST && cType.id() == IType.LIST) {
 			if (exp instanceof ListExpression) {
 				final IExpression[] array = ((ListExpression) exp).getElements();
-				if (array.length == 0) {
-					return Types.NO_TYPE;
-				}
+				if (array.length == 0) { return Types.NO_TYPE; }
 				return array[0].getType().getContentType();
 			} else if (exp instanceof MapExpression) {
 				final IExpression[] array = ((MapExpression) exp).valuesArray();
-				if (array.length == 0) {
-					return Types.NO_TYPE;
-				}
+				if (array.length == 0) { return Types.NO_TYPE; }
 				return array[0].getType().getContentType();
 			} else {
 				return cType.getContentType();
 			}
 		}
-		if (Types.CONTAINER.isAssignableFrom(itemType)) {
-			return itemType.getContentType();
-		}
+		if (Types.CONTAINER.isAssignableFrom(itemType)) { return itemType.getContentType(); }
 		return itemType;
 	}
